Character seeds now use ulong instead of uint
That actually wasn't hard at all!

+++ Assets/Source/Gameplay/Player/BotPlayer.cs
@@ -32,7 +32,7 @@ public class BotPlayer : Player
 
     }
 
-    public override void guessTarget(uint characterId)
+    public override void guessTarget(ulong characterId)
     {
         bool correct = TurnDriver.instance.currentRoster.targetId == characterId;
